requirements.txt
PyYAML
# prometheus_client
statsd
requests